Video Script, Can you follow me?
Please just note that the extra features may need small additions, and I've tried to give generous amounts of time towards the end (the timestamps seem to finish at the best part of 8-9 mins rather than 10). This is to give me the best possible opportunity to fit in any displaying of the features - though that could be sped up, we still need to be able to understand what is going on. Can you follow me?